+++ packages/core/src/index.ts
@@ -5,11 +5,3 @@ export * from "./generators/css-classes";
 export * from "./generators/tokens";
 export * from "./generators/score";
 export * from "./generators/formats";
-
-export * from "./hooks/usePerformanceMetrics";
-export * from "./hooks/useTimeBeforeInteraction";
-export * from "./hooks/useComponentsInteractions";
-export * from "./hooks/useABTesting";
-export * from "./hooks/useBehaviorAnalytics";
-export * from "./hooks/useIndexedDB";
-export * from "./hooks/useAppProperties";