+++ packages/hooks/src/index.ts
@@ -1,7 +1,4 @@
 export * from "./useClickOutside";
 export * from "./useEventListener";
 export * from "./useKeyPress";
-
-export * from "./useTimer";
-export * from "./useSaveOnUnload";
 export * from "./useDisabledScroll";