+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @foundation-ui/core
 
+## 1.3.1
+
+### Patch Changes
+
+- Add @fui/icon, @fui/analytics; Update name in License
+- Updated dependencies
+  - @foundation-ui/hooks@1.3.1
+
 ## 1.3.0
 
 ### Minor Changes